chore: move css closer to usage, and delete unused files
