+++ src/Altinn.Apps/AppFrontend/react/altinn-app-frontend/src/components/base/FileUpload/FileUploadComponent.css
