+++ src/Altinn.Apps/AppFrontend/react/altinn-app-frontend/src/index.css
