--- src/Altinn.Apps/AppFrontend/react/altinn-app-frontend/src/styles/ErrorMessageComponent.css
@@ -1,44 +1,0 @@
-.error-snackbar-container {
-  min-width: 250px;
-  margin-left: -125px;
-  position: fixed;
-  bottom: 30px;
-  left: 50%;
-  z-index: 1000;
-}
-
-.error-snackbar-items {
-  margin: 5px;
-  background-color: red;
-  border: 2px rgba(0, 0, 0, 0);
-  color: white;
-  text-align: center;
-  padding: 5px;
-  top: 2px;
-  font-size: 17px;
-  z-index: 999999999999999999999999999999999;
-  -webkit-animation: fadein 0.5s;
-  animation: fadein 0.5s;
-}
-
-@-webkit-keyframes fadein {
-  from {
-    bottom: 0;
-    opacity: 0;
-  }
-  to {
-    bottom: 30px;
-    opacity: 1;
-  }
-}
-
-@keyframes fadein {
-  from {
-    bottom: 0;
-    opacity: 0;
-  }
-  to {
-    bottom: 30px;
-    opacity: 1;
-  }
-}
