--- src/Altinn.Apps/AppFrontend/react/altinn-app-frontend/src/styles/toolBar.css
@@ -1,20 +1,0 @@
-.a-item {
-  display: flex;
-  user-select: none;
-  padding: 0.5rem !important;
-  margin: 0 0 0.5rem 0 !important;
-  align-items: flex-start;
-  align-content: flex-start;
-  line-height: 1.5;
-  border-radius: 3px;
-  background: #fff;
-  border: 1px solid #ddd;
-}
-
-.a-add {
-  /* padding-left: 2.5rem; */
-}
-
-.a-item+.a-item-clone {
-  display: none !important;
-}
